+++ Sources/TiledKit/Coding/XML/XMLProperties+PropertyValue.swift
